fixing broken tests

+++ tests/customed_header.h
@@ -1,5 +1,5 @@
 
-static inline void func(CField *f, float *lon, float *dt)
+static inline void func(CField *f, double *lon, double *dt)
 {
   float (*data)[f->xdim] = (float (*)[f->xdim]) f->data;
   float u = data[2][1];
